feat(autocomplete): add standalone AutocompleteController and consolidate routes
Add AutocompleteController as the new generic autocomplete endpoint,
supporting both AutocompleterInterface and EntityAutocompleterInterface
(for backwards compatibility).

- Rename route ux_entity_autocomplete -> ux_autocomplete
- Add ux_entity_autocomplete as a deprecated route alias
- Update AutocompleteResultsExecutor to dispatch to AutocompleterInterface
  directly when the autocompleter does not require Doctrine
- Update AutocompleterRegistry return type to accept AutocompleterInterface
- Deprecate EntityAutocompleteController in favor of AutocompleteController
